% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_treemap.R
\name{draw_key_rrect}
\alias{draw_key_rrect}
\title{Round rect key glyph for legend}
\usage{
draw_key_rrect(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\description{
Round rect key glyph for legend
}
\author{
Bob Rudis (bob@rud.is)
}
