% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{G20}
\alias{G20}
\title{Statistics on the G-20 group of major world economies.}
\format{A data frame with 20 rows and five variables:
\describe{
  \item{region}{the country's region}
  \item{country}{the country}
  \item{gdp_mil_usd}{the country's GDP, in millions of US dollars}
  \item{hdi}{the country's Human Development Index}
  \item{econ_classification}{the country's economic classification}
}}
\source{
\url{http://en.wikipedia.org/wiki/G-20_major_economies}
}
\usage{
G20
}
\description{
A dataset containing economic and demographic statistics about members of the
G-20 group of major world economies.
}
\keyword{datasets}
