% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_treemap_group_text.R
\name{geom_treemap_group_text}
\alias{geom_treemap_group_text}
\title{Text labels for treemap groups.}
\usage{
geom_treemap_group_text(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{Standard
geom arguments as for \code{geom_rect}.}

\item{padding.x, padding, y}{Unit object, giving horizontal or vertical padding
between text and edge of tile. Defaults to 1 mm.}

\item{place}{Where should the text be drawn within the group? One of
‘topleft’, ‘top’, ‘topright’ etc. Defaults to ‘bottom’.}

\item{min.size}{Number, in points. Text that would need to be drawn smaller
than this size to fit in the group will be hidden. Defaults to 4 pt.}

\item{fill.text}{Logical; should text be expanded to fill the entire group?
Defaults to false.}
}
\description{
Requires ‘area’ and ‘group’ ‘group.label’ aesthetics. Strange things will
happen if ‘group’ and ‘group.label’ don't carve the data into the same
groups, or if \code{geom_treemap_group_text} is given a different dataset or
area aesthetic than the \code{geom_treemap} it is drawn over.

\code{geom_treemap_group_text} uses \code{geom_fit_text} from the
\code{ggfittext} package to fit text to the group All text drawing options
available in \code{ggfittext} (e.g. expanding text to fill a group) are also
available here.
}
\section{Aesthetics}{


\itemize{
  \item area (required)
  \item label (required)
  \item colour
  \item size
  \item alpha
  \item family
  \item fontface
  \item angle
}
}
\seealso{
geom_treemap, geom_treemap_group_border
}

