% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_treemap_subgroup_border.R
\name{geom_treemap_subgroup_border}
\alias{geom_treemap_subgroup_border}
\title{Draw a border around a subgroup of treemap tiles.}
\usage{
geom_treemap_subgroup_border(mapping = NULL, data = NULL,
  stat = "identity", position = "identity", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{Standard
geom arguments as for \code{geom_rect}.}
}
\description{
Requires ‘area’ and ‘subgroup’. Strange things will happen if
\code{geom_treemap_subgroup_border} is given a different dataset or area
aesthetic than the \code{geom_treemap} it is drawn over.
}
\section{Aesthetics}{


\itemize{
  \item area (required)
  \item subgroup (required)
  \item colour
  \item size
  \item linetype
  \item alpha
}
}
\seealso{
geom_treemap, geom_treemap_subgroup_text
}

