% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_treemap.R, R/geom_treemap_group_border.R, R/geom_treemap_group_text.R, R/geom_treemap_text.R
\docType{data}
\name{GeomTreemap}
\alias{GeomTreemap}
\alias{GeomTreemapSubgroupBorder}
\alias{GeomTreemapSubgroupText}
\alias{GeomTreemapText}
\title{GeomTreemap}
\description{
GeomTreemap
}
\keyword{datasets}

