% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplotify.R
\name{ggplotify}
\alias{ggplotify}
\title{Draw an exploratory treemap}
\usage{
ggplotify(treeMap, label.groups = TRUE, label.colour = "white",
  group.label.colour = "darkgrey")
}
\arguments{
\item{treeMap}{a data frame of treemap coordinates produced by "treemapify"}

\item{label.groups}{should groups be labeled? (Individual observations will
be automatically labelled if a "label" parameter was passed to
"treemapify")}

\item{label.colour}{colour for individual rect labels; defaults to white}

\item{group.label.colour}{colour for group labels; defaults to darkgrey}
}
\description{
Takes a data frame of treemap coordinates produced by "treemapify" and
draws an exploratory treemap.  The output is a ggplot2 plot, so it can be
further manipulated e.g. a title added.
}
\seealso{
Other treemapify: \code{\link{treemapify}}
}

