% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_treemap.R, R/geom_treemap_text.R,
%   R/utils.R
\docType{data}
\name{GeomTreemap}
\alias{GeomTreemap}
\alias{GeomTreemapText}
\alias{treemapify-ggproto}
\title{GeomTreemap}
\description{
GeomTreemap

GeomTreemapText

Base ggproto classes for treemapify extension to ggplot2
}
\seealso{
ggplot2::ggproto
}
\keyword{datasets}
\keyword{internal}
