% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemapify-package.R
\docType{package}
\name{treemapify-package}
\alias{treemapify-package}
\title{treemapify: Draw Treemaps in 'ggplot2'}
\description{
Provides 'ggplot2' geoms for drawing treemaps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wilkox.org/treemapify/}
  \item Report bugs at \url{https://github.com/wilkox/treemapify/issues/}
}

}
\author{
\strong{Maintainer}: David Wilkins \email{david@wilkox.org} (\href{https://orcid.org/0000-0002-3385-8981}{ORCID})

Other contributors:
\itemize{
  \item Bob Rudis \email{bob@rud.is} (\href{https://orcid.org/0000-0001-5670-2640}{ORCID}) [contributor]
}

}
\keyword{internal}
