% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_treemap.R
\name{geom_treemap}
\alias{geom_treemap}
\title{A treemap}
\usage{
geom_treemap(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{Standard
geom arguments as for \code{geom_rect}.}
}
\description{
Draw a treemap, where each observation is represented by a tile scaled by an
area aesthetic.

‘area’ is the only required aesthetic. Other aesthetics are accepted as for
\code{geom_rect}. See \code{geom_treemap_text} to label tiles.

An optional ‘subgroup’ aesthetic will cause the tiles to be clustered in
subgroups within the treemap. See \code{geom_treemap_subgroup_border} and
\code{geom_treemap_subgroup_text} to draw a border around subgroups and label
them respectively.

Tile placement proceeds from the bottom left corner, moving alternately
rightwards and upwards until all tiles are placed. See Bruls et al. (1999)
for the full algorithm.
}
\section{Aesthetics}{


\itemize{
  \item area (required)
  \item alpha
  \item colour
  \item fill
  \item linetype
  \item subgroup
}
}
\references{
treemapify uses the Squarified Treemap algorithm of Mark Bruls, Kees Huizing
and Jarke van Wijk:

Bruls, M., Huizing, K., & van Wijk, J. (1999). Squarified Treemaps (pp.
33-42). Presented at the In Proceedings of the Joint Eurographics and IEEE
TCVG Symposium on Visualization.
\url{http://www.win.tue.nl/~vanwijk/stm.pdf}
}
\seealso{
geom_treemap_text, geom_treemap_subgroup_border,
geom_treemap_subgroup_text
}

