% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_treemap_text.R
\name{geom_treemap_text}
\alias{geom_treemap_text}
\title{Add text labels to treemap tiles.}
\usage{
geom_treemap_text(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = FALSE,
  inherit.aes = TRUE, padding.x = unit(1, "mm"), padding.y = unit(1,
  "mm"), place = "topleft", min.size = 4, grow = F, ...)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{Standard
geom arguments as for \code{geom_rect}.}

\item{padding.x, padding, y}{Unit object, giving horizontal or vertical padding
between text and edge of tile. Defaults to 1 mm.}

\item{place}{Where should the text be drawn within the tile? One of
‘topleft’, ‘top’, ‘topright’ etc. Defaults to ‘topleft’.}

\item{min.size}{Number, in points. Text that would need to be drawn smaller
than this size to fit in the tile will be hidden. Defaults to 4 pt.}

\item{grown}{Logical; should text grow to fill the entire tile?
Defaults to false.}
}
\description{
This will draw a text label for each tile in a treemap created by
\code{geom_treemap}. In addition to the required ‘label’ aesthetic, it
requires an ‘area’ aesthetic as used in \code{geom_treemap}; if this is
mapped to a different column, or \code{geom_treemap} and
\code{geom_treemap_text} are passed different data, the text labels probably
won't line up with the tiles.

\code{geom_treemap_text} uses \code{geom_fit_text} from the \code{ggfittext}
package to fit text to the tile. All text drawing options available in
\code{ggfittext} (e.g. growing text to fill a tile) are also available here.
}
\section{Aesthetics}{


\itemize{
  \item area (required)
  \item label (required)
  \item colour
  \item size
  \item alpha
  \item family
  \item fontface
  \item angle
}
}
\seealso{
geom_treemap, geom_fit_text
}

